funky rounding loan created

+++ lib/functions.rb
@@ -157,7 +157,7 @@ class String
 end
 
 
-class Float
+class Numeric
   alias_method :round_orig, :round
   def round(n=0)
     (self * (10.0 ** n)).round_orig * (10.0 ** (-n))
